CAE support in VMwrtbarWithoutStoreEvaluator
VMwrtbarWithoutStoreEvaluator is used for several different opcdes include
arraycopy, ArrayStoreCHK, writeBarrier and Unsafe CAS calls. This is only used
for the Unsafe CAS calls that store an object.

To differtiate Unsafe CAS calls from the other cases, there is a check for the
node being an icall. This check is no longer valid with the introduction of
support for compareAndExchangeReference. The node in this case is an acall.

This changes the check to a generic call check. There is also a check for the
call being to a recognized method. This check now checks for specific
recognized methods which are:

sun_misc_Unsafe_compareAndSwapObject_jlObjectJjlObjectjlObject_Z
jdk_internal_misc_Unsafe_compareAndExchangeObject
jdk_internal_misc_Unsafe_compareAndExchangeReference